Build fix for the mac in the new file transfer code.
2007-06-04  Mikael Hallendal  <micke@imendio.com>

	* libloudermouth/lm-bs-receiver.c (bs_receiver_fill_request_buf):
	* libloudermouth/lm-bs-sender.c (bs_sender_fill_request_buf):
	- Cast return value from strlen to int to build on mac.

svn path=/trunk/; revision=2499

+++ libloudermouth/lm-bs-receiver.c
@@ -102,7 +102,7 @@ bs_receiver_fill_request_buf (LmBsClient   *client,
 {
 	*buf = g_string_new_len ("\x05\x01\x00\x03", 4);
 
-	g_string_append_printf (*buf, "%c%s", strlen (msg), msg);
+	g_string_append_printf (*buf, "%c%s", (int)strlen (msg), msg);
 	g_string_append_len (*buf, "\x00\x00", 2);
 }
